+++ src/lightning_ml/core/__init__.py
@@ -2,10 +2,4 @@
 
 from __future__ import annotations
 
-# from .data.datamodule import DataModule
-from . import data
-from .abstract.learner import Learner
-from .abstract.predictor import Predictor
-from .abstract.school import School
-
-__all__ = ["Learner", "Predictor", "School"]
+from . import abstract, data, learners, predictors, utils
