+++ src/lightning_ml/core/abstract/__init__.py
@@ -1,5 +1,6 @@
 """Abstract base classes for Lightning-ML."""
 
+from .dataset import BaseDataset
 from .learner import Learner
 from .predictor import Predictor
 from .school import School
