+++ src/lightning_ml/core/abstract/dataset.py
@@ -8,39 +8,14 @@ This module defines:
 """
 
 from abc import ABCMeta, abstractmethod
-
-# Copyright The PyTorch Lightning team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any
 
 import torch
 from torch.utils.data import Dataset
 
-from ..utils.enums import DataKeys
 from .sample import Sample
 
-__all__ = [
-    "BaseDataset",
-    "InputMixin",
-    "TargetMixin",
-    "PositiveMixin",
-    "NegativeMixin",
-    "UnlabelledDatasetBase",
-    "LabelledDatasetBase",
-    "ContastiveDatasetBase",
-    "TripletDatasetBase",
-]
+__all__ = ["BaseDataset"]
 
 
 class DatasetMeta(ABCMeta):
@@ -140,65 +115,3 @@ class BaseDataset(Dataset, metaclass=DatasetMeta):
         Returns:
             int: Number of samples.
         """
-
-
-class InputMixin(BaseDataset):
-    """Dataset mix-in that mandates a `get_input` method, the output of which
-    will be added to the `__getitem__` ouput dict of all subclasses with key
-    "input".
-    """
-
-    sample_keys = [DataKeys.INPUT]
-
-
-class TargetMixin(BaseDataset):
-    """Dataset mix-in that mandates a `get_target` method."""
-
-    sample_keys = [DataKeys.TARGET]
-
-
-class PositiveMixin(BaseDataset):
-    """Dataset mix-in that mandates a `get_positive` method."""
-
-    sample_keys = [DataKeys.POSITIVE]
-
-
-class NegativeMixin(BaseDataset):
-    """Dataset mix-in that mandates a `get_negative` method."""
-
-    sample_keys = [DataKeys.NEGATIVE]
-
-
-# Abstract dataset bases
-
-
-class UnlabelledDatasetBase(InputMixin):
-    """
-    Abstract base for unlabeled datasets.
-
-    Combines mix-ins to enforce retrieval of input samples only.
-    """
-
-
-class LabelledDatasetBase(InputMixin, TargetMixin):
-    """
-    Abstract base for labeled datasets.
-
-    Combines mix-ins to enforce retrieval of input and target samples.
-    """
-
-
-class ContastiveDatasetBase(InputMixin, PositiveMixin):
-    """
-    Abstract base for contrastive datasets.
-
-    Combines mix-ins to enforce retrieval of input and positive samples.
-    """
-
-
-class TripletDatasetBase(ContastiveDatasetBase, NegativeMixin):
-    """
-    Abstract base for triplet datasets.
-
-    Combines mix-ins to enforce retrieval of input, positive, and negative samples.
-    """
