--- src/lightning_ml/core/data/dataset.py
@@ -1,3 +1,0 @@
-"""Compatibility wrapper forwarding to :mod:`lightning_ml.core.abstract.dataset`."""
-
-from ..abstract.dataset import *  # noqa: F401,F403
