+++ src/lightning_ml/core/data/datasets/__init__.py
@@ -14,4 +14,3 @@ REGISTRY = get_registry(Registries.DATASET)
 from .contrastive import *  # noqa: F401,F403
 from .labelled import *  # noqa: F401,F403
 from .unlabelled import *  # noqa: F401,F403
-from .numpy import *  # noqa: F401,F403
