--- src/lightning_ml/core/data/loader.py
@@ -1,3 +1,0 @@
-"""Compatibility wrapper forwarding to :mod:`lightning_ml.core.abstract.loader`."""
-
-from ..abstract.loader import *  # noqa: F401,F403
