--- src/lightning_ml/core/data/sample.py
@@ -1,3 +1,0 @@
-"""Compatibility wrapper forwarding to :mod:`lightning_ml.core.abstract.sample`."""
-
-from ..abstract.sample import *  # noqa: F401,F403
